+++ src/app/loading.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const loading = () => {
+	return <div>loading</div>;
+};
+
+export default loading;
