+++ src/middleware.ts
@@ -1,12 +1,11 @@
-import NextAuth from "next-auth";
-
-import type { NextRequest } from "next/server";
-import { NextResponse } from "next/server";
+import { NextResponse, type NextRequest } from "next/server";
 
 export async function middleware(request: NextRequest) {
 	const headers = new Headers(request.headers);
 	headers.set("x-current-path", request.nextUrl.pathname);
 
+	// TODO: add the logic to redirect user to /create if loggedin and want to access the /login, /signin or /signup
+	// TODO: to loggin if the user is not logged in and want to access the /profile or /create
 	return NextResponse.next({ headers });
 }
